requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
anaconda-anon-usage @ file:///opt/miniconda3/conda-bld/anaconda-anon-usage_1764636648062/work
anaconda-auth @ file:///C:/miniconda3/conda-bld/anaconda-cloud-auth-split_1765839853030/work
anaconda-cli-base @ file:///C:/miniconda3/conda-bld/anaconda-cli-base_1764888883727/work
annotated-types==0.7.0
anyio==4.9.0
archspec @ file:///home/task_175812491784513/conda-bld/archspec_1758124989039/work
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boltons @ file:///C:/b/abs_e2_iokhxbp/croot/boltons_1751383740243/work
brotlicffi @ file:///C:/miniconda3/conda-bld/brotlicffi_1764961374486/work
cachetools==6.1.0
certifi==2025.6.15
cffi @ file:///C:/miniconda3/conda-bld/cffi_1761832792955/work
charset-normalizer==3.4.2
click==8.2.1
click-default-group==1.2.4
colorama==0.4.6
conda @ file:///C:/miniconda3/conda-bld/conda_1765567840841/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1755883788794/work
conda-anaconda-tos @ file:///C:/b/abs_c9yejtsx9g/croot/conda-anaconda-tos_1755123332641/work
conda-content-trust @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/conda-content-trust_1729088072778/work
conda-libmamba-solver @ file:///C:/miniconda3/conda-bld/conda-libmamba-solver_1764245615235/work/src
conda-package-handling @ file:///C:/miniconda3/conda-bld/conda-package-handling_1762366515145/work
conda_package_streaming @ file:///C:/miniconda3/conda-bld/conda-package-streaming_1762361689679/work
cryptography @ file:///C:/miniconda3/conda-bld/cryptography-split_1761932023265/work
diffusers==0.34.0
distro==1.9.0
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.1
frozendict @ file:///C:/miniconda3/conda-bld/frozendict_1761750728192/work
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google_search_results==2.4.2
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.3.0
itsdangerous==2.2.0
jaraco.classes @ file:///C:/b/abs_6erueoob1v/croot/jaraco.classes_1755516340851/work
jaraco.context @ file:///C:/Users/dev-admin/buildout/perseverance-python-buildout/croot/jaraco.context_1731721000658/work
jaraco.functools @ file:///C:/b/abs_a1jv6v_pzp/croot/jaraco.functools_1740408129620/work
Jinja2==3.1.6
jiter==0.10.0
jsonpatch @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/jsonpatch_1729054776004/work
jsonpointer @ file:///C:/b/abs_73u73l7pl9/croot/jsonpointer_1753788460913/work
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keyring @ file:///C:/miniconda3/conda-bld/keyring_1763637203252/work
libmambapy @ file:///C:/miniconda3/conda-bld/mamba-split_1763111608356/work/libmambapy
markdown-it-py @ file:///C:/b/abs_3espcrdl35/croot/markdown-it-py_1756299213703/work
MarkupSafe==3.0.2
mdurl @ file:///C:/miniconda3/conda-bld/mdurl_1758552277768/work
menuinst @ file:///C:/miniconda3/conda-bld/menuinst_1765382397455/work
more-itertools @ file:///C:/miniconda3/conda-bld/more-itertools_1761121564395/work
mpmath==1.3.0
msgpack @ file:///C:/b/abs_4b3t4uhz3r/croot/msgpack-python_1750958631084/work
multidict==6.5.0
narwhals==1.43.1
networkx==3.5
numpy==2.2.5
openai==1.88.0
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pkce @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pkce_1729049216383/work
platformdirs @ file:///C:/miniconda3/conda-bld/platformdirs_1762356623609/work
playwright==1.52.0
pluggy==1.6.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pycosat @ file:///C:/b/abs_18nblzzn70/croot/pycosat_1736868434419/work
pycparser @ file:///C:/miniconda3/conda-bld/pycparser_1757496153123/work
pydantic==2.11.7
pydantic-settings @ file:///C:/miniconda3/conda-bld/pydantic-settings_1764165236385/work
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments @ file:///C:/miniconda3/conda-bld/pygments_1762431428918/work
PyJWT @ file:///C:/miniconda3/conda-bld/pyjwt_1764332257117/work
PySocks @ file:///C:/miniconda3/conda-bld/pysocks_1761753030965/work
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32-ctypes @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pywin32-ctypes_1729046491215/work
PyYAML==6.0.2
qrcode==8.1
readchar @ file:///C:/miniconda3/conda-bld/readchar_1760613474723/work
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich @ file:///C:/miniconda3/conda-bld/rich_1760375661587/work
rpds-py==0.25.1
ruamel.yaml @ file:///C:/miniconda3/conda-bld/ruamel.yaml_1762536064547/work
ruamel.yaml.clib @ file:///C:/miniconda3/conda-bld/ruamel.yaml.clib_1762530094515/work
safetensors==0.5.3
scipy==1.17.1
semver @ file:///C:/miniconda3/conda-bld/semver_1761903323755/work
setuptools==80.9.0
shellingham @ file:///C:/miniconda3/conda-bld/shellingham_1761912227081/work
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlite-fts4==1.0.3
sqlite-utils==3.38
streamlit==1.46.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
tomli @ file:///C:/b/abs_88e598m6o8/croot/tomli_1753774604115/work
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
truststore @ file:///C:/miniconda3/conda-bld/truststore_1762521027919/work
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
win_inet_pton @ file:///C:/miniconda3/conda-bld/win_inet_pton_1761746278300/work
yarl==1.20.1
zipp==3.23.0
zstandard @ file:///C:/miniconda3/conda-bld/zstandard_1758189089298/work
